+++ tests/test_09_deprecations.py
@@ -4,4 +4,9 @@
 import pytest
 import icat
 
-# No deprecations for the moment.
+# Deprecations not tested in this module:
+# - Module variable icat.config.defaultsection.
+#   It's easier to test this in the setting of the test_01_config.py
+#   module.
+#
+# No other deprecations for the moment.